Refactor ZwiftMemoryMonitor and related classes
- Moved ZwiftMemoryMonitor implementation to a separate file for better organization.
- Introduced ZwiftProcess class to handle process-related operations, improving separation of concerns.
- Updated ZwiftMemoryScanner to utilize the new ZwiftProcess class for process management.
- Changed method signatures in ZwiftMemoryScanner to use async/await for better asynchronous handling.
- Updated example state.js to demonstrate new functionality.
- Cleaned up code and improved logging for better debugging and error handling.
