+++ package.json
@@ -35,9 +35,11 @@
     "node": ">=18"
   },
   "dependencies": {
+    "@zwfthcks/zwift-data": "github:zwfthcks/zwift-data",
     "memoryjs": "github:zwfthcks/memoryjs",
     "platform-paths": "^1.2.2",
-    "semver": "^7.7.2"
+    "semver": "^7.7.2",
+    "zwift-data": "github:zwfthcks/zwift-data"
   },
   "devDependencies": {
     "ansi-escapes": "4.3.2",
