--- src/AddressCache.js
@@ -1,6 +1,0 @@
-
-
-class AddressCache {
-
-    // use a Map to store the address cache
-}
