+++ src/examples/state.js
@@ -71,11 +71,11 @@ zmm.once('ready', () => {
     try {
         // zmm.start(['playerstate'])
         // zmm.start(['playerstate', 'playerprofile'])
-        // zmm.start(['playerstate', 'playerprofile'], { forceScan: true })
+        zmm.start(['playerstate', 'playerprofile'], { forceScan: true })
         // zmm.start(['playerstate'], { timeout: 500,})
         // zmm.start(['playerstate'], { forceScan: true, multiPass: true,})
         // zmm.start(['playerstate'], { forceScan: true})
-        zmm.start(['playerprofile'], { forceScan: true })
+        // zmm.start(['playerprofile'], { forceScan: true })
         // zmm.start(['playerprofile'], { forceScan: true, multiPass: true })
         console.log('>>', 'last error:', zmm.lasterror)
     } catch (e) {
